requirements.txt
python-docx
simplify_docx==0.1.2
